Fix problems

+++ core-c/src/c/services/utils.c
@@ -17,7 +17,7 @@
 #include "services/utils.h"
 
 void buffer_message_set_size(buffer *b) {
-    *((short *) b->start) = b->temp - b->start;
+    *((short *) b->start) = (short) ((b->temp) - (b->start));
 }
 
 void free_uint32(void *v) {
